requirements.txt
amqp==5.0.9
asgiref==3.5.0
attrs==21.4.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
boto3==1.21.4
botocore==1.24.4
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
defusedxml==0.7.1
Django==4.0.2
django-cors-headers==3.11.0
django-environ==0.8.1
django-rest-multiple-models==2.1.3
django-storages==1.12.3
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-spectacular==0.21.2
idna==3.3
importlib-resources==5.4.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jmespath==0.10.0
jsonschema==4.4.0
kombu==5.2.3
MarkupSafe==2.0.1
oauthlib==3.2.0
Pillow==9.0.1
prompt-toolkit==3.0.27
pycparser==2.21
PyJWT==2.3.0
pyrsistent==0.18.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
s3transfer==0.5.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
zipp==3.7.0
